# portrait-studio/docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "GenerateRequest": {
        "additionalProperties": false,
        "properties": {
          "adult_only": {
            "default": false,
            "title": "Adult Only",
            "type": "boolean"
          },
          "allow_locked_override": {
            "default": false,
            "title": "Allow Locked Override",
            "type": "boolean"
          },
          "cfg": {
            "anyOf": [
              {
                "maximum": 30.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cfg"
          },
          "character_id": {
            "title": "Character Id",
            "type": "string"
          },
          "count": {
            "default": 4,
            "maximum": 8.0,
            "minimum": 1.0,
            "title": "Count",
            "type": "integer"
          },
          "face_detailer": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Face Detailer"
          },
          "height": {
            "anyOf": [
              {
                "maximum": 2048.0,
                "minimum": 512.0,
                "multipleOf": 8.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "negative_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Negative Prompt"
          },
          "overrides": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LockedParamsPatch"
              },
              {
                "type": "null"
              }
            ]
          },
          "prompt": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          },
          "sampler_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sampler Name"
          },
          "save_as_version": {
            "default": false,
            "title": "Save As Version",
            "type": "boolean"
          },
          "scene_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 5,
            "title": "Scene Ids",
            "type": "array"
          },
          "scheduler": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduler"
          },
          "seed": {
            "default": -1,
            "minimum": -1.0,
            "title": "Seed",
            "type": "integer"
          },
          "steps": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Steps"
          },
          "upscale": {
            "anyOf": [
              {
                "maximum": 2.0,
                "minimum": 1.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upscale"
          },
          "width": {
            "anyOf": [
              {
                "maximum": 2048.0,
                "minimum": 512.0,
                "multipleOf": 8.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          }
        },
        "required": [
          "character_id",
          "prompt"
        ],
        "title": "GenerateRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobAccepted": {
        "additionalProperties": false,
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          }
        },
        "required": [
          "job_id",
          "position"
        ],
        "title": "JobAccepted",
        "type": "object"
      },
...
      "LockedParamsPatch": {
        "additionalProperties": false,
        "description": "Partial LockedParams (used for `overrides`, registration adjustments and new versions).",
        "properties": {
          "checkpoint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkpoint"
          },
          "default_height": {
            "anyOf": [
              {
                "maximum": 2048.0,
                "minimum": 512.0,
                "multipleOf": 8.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Height"
          },
          "default_width": {
            "anyOf": [
              {
                "maximum": 2048.0,
                "minimum": 512.0,
                "multipleOf": 8.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Width"
          },
          "face_detailer_denoise": {
            "anyOf": [
              {
                "maximum": 0.49,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Face Detailer Denoise"
          },
          "face_method": {
            "anyOf": [
              {
                "enum": [
                  "pulid",
                  "faceid",
                  "instantid"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Face Method"
          },
          "face_weight": {
            "anyOf": [
              {
                "maximum": 1.5,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Face Weight"
          },
          "hires_max": {
            "anyOf": [
              {
                "maximum": 2.0,
                "minimum": 1.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hires Max"
          },
          "lora": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lora"
          },
          "lora_strength": {
            "anyOf": [
              {
                "maximum": 2.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lora Strength"
          },
          "negative_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Negative Prompt"
          },
          "prefix_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prefix Prompt"
          }
        },
        "title": "LockedParamsPatch",
        "type": "object"
      },
...
      "PreviewVramRequest": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "default": 1,
            "maximum": 8.0,
            "minimum": 1.0,
            "title": "Count",
            "type": "integer"
          },
          "face_detailer": {
            "default": true,
            "title": "Face Detailer",
            "type": "boolean"
          },
          "height": {
            "default": 1216,
            "maximum": 8192.0,
            "minimum": 64.0,
            "title": "Height",
            "type": "integer"
          },
          "method": {
            "default": "pulid",
            "title": "Method",
            "type": "string"
          },
          "upscale": {
            "default": 1.0,
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Upscale",
            "type": "number"
          },
          "width": {
            "default": 832,
            "maximum": 8192.0,
            "minimum": 64.0,
            "title": "Width",
            "type": "integer"
          }
        },
        "title": "PreviewVramRequest",
        "type": "object"
      },
      "PreviewVramResponse": {
        "additionalProperties": false,
        "properties": {
          "advice": {
            "title": "Advice",
            "type": "string"
          },
          "basis": {
            "enum": [
              "measured",
              "bounded",
              "unknown"
            ],
            "title": "Basis",
            "type": "string"
          },
          "estimated_peak_mb": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Peak Mb"
          },
          "free_mb": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Free Mb"
          },
          "risk": {
            "enum": [
              "low",
              "medium",
              "high",
              "unknown"
            ],
            "title": "Risk",
            "type": "string"
          },
          "total_mb": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Mb"
          },
          "would_reject": {
            "title": "Would Reject",
            "type": "boolean"
          }
        },
        "required": [
          "estimated_peak_mb",
          "free_mb",
          "total_mb",
          "risk",
          "basis",
          "advice",
          "would_reject"
        ],
        "title": "PreviewVramResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "同一の架空キャラクター（実写調・日本人・成人）で画像を生成・管理するローカル専用 API。認証: ヘッダ `X-API-Key`（`GET /api/health` を除く）。エラーは `{\"detail\": \"...\"}`（日本語）。",
    "title": "portrait-studio API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/generate": {
      "post": {
        "operationId": "generate_api_generate_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate",
        "tags": [
          "generate"
        ]
      }
    },
    "/api/generate/preview-vram": {
      "post": {
        "operationId": "preview_vram_api_generate_preview_vram_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewVramRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewVramResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Vram",
        "tags": [
          "generate"
        ]
      }
    },
...
  }
}